% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2-package.R
\docType{package}
\name{co2meter}
\alias{co2meter}
\alias{co2meter-package}
\title{Retrieve Readings from CO2Mini (Model RAD-0301) Desktop CO2 Monitor}
\description{
The CO2Mini (model RAD-0301) desktop CO2 monitor from CO2meter.com
is a low-cost, fairly accurate personal CO2 environment monitor that
can work standalone (via a USB power brick) or be attacked to a computer
where it can be both powered and act as a datalogger. Tools are provided
to retrieve temperature and CO2 readings from such devices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/co2meter}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/co2meter/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
