% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{read_co2}
\alias{read_co2}
\alias{close_dev}
\alias{init_co2_meter_dev}
\title{Initialize, read from, and close a co2meter.com device}
\usage{
read_co2()

close_dev()

init_co2_meter_dev()
}
\value{
\code{\link[=get_readings]{get_readings()}} returns a named list containing ambient temperature (°C) and CO2 (ppm)
}
\description{
Initialize, read from, and close a co2meter.com device
}
\examples{
\dontrun{
init_co2_meter_dev()
read_co2()
close_dev()
}
}
